Handle network errors by returning error result
